requirements.txt
langchain
langgraph
langchain-ollama
gradio
faiss-cpu
sentence-transformers
matplotlib